shares with minimun test OK

+++ contracts/ERC223BasicToken.sol
@@ -26,10 +26,10 @@ contract ERC223BasicToken is ERC223Basic, BasicToken {
      * @param _value Amount of tokens that will be transferred.
      * @param _data  Transaction metadata.
      */
-    function transfer(address _to, uint _value, bytes _data) public returns (bool) {
+    function transfer(address _to, uint256 _value, bytes _data) public returns (bool) {
         // Standard function transfer similar to ERC20 transfer with no _data .
         // Added due to backwards compatibility reasons .
-        uint codeLength;
+        uint256 codeLength;
         assembly {
             // Retrieve the size of the code on target address, this needs assembly .
             codeLength := extcodesize(_to)
